Setup callbacks before initializing 3D Graphics
The intent is to capture initial window resize events from a tiling
window manager. Otherwise there is no way to find the initial window size.
Samples presume 800x600, which works in a non-tiling window manager.

To actually work, this probably needs a delay or some kind of synchronization
between threads. This initial patch is an attempt to just figure out how
to build HGamer3D and call the modified version, and see if changing the
order of initialization breaks anything.

Another solution would be to expose an API to get the window size,
so the application can request it on demand, instead of assuming
the initial window size matches the requested window size.